Added javadoc, did some minor renaming
