+++ maven-surefire-common/src/main/java/org/apache/maven/plugin/surefire/booterclient/lazytestprovider/FlushReceiver.java
@@ -1,7 +1,37 @@
 package org.apache.maven.plugin.surefire.booterclient.lazytestprovider;
 
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
 import java.io.IOException;
 
+/**
+ * Something that can be flushed.
+ * 
+ * @author Andreas Gudian
+ *
+ */
 public interface FlushReceiver {
+	/**
+	 * Performs a flush, releasing any buffered resources.
+	 * 
+	 * @throws IOException in case the flush operation failed
+	 */
 	void flush() throws IOException;
 }
