+++ surefire-api/src/main/java/org/apache/maven/surefire/util/LazyTestsToRun.java
@@ -1,7 +1,23 @@
-/**
- * 
- */
 package org.apache.maven.surefire.util;
+
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -16,7 +32,7 @@ import java.util.List;
 import org.apache.maven.surefire.booter.ForkingRunListener;
 
 /**
- * A variant of TestsToRun that is provided with test class names asynchronously
+ * A variant of TestsToRun that is provided with test class names
  * from an {@link InputStream} (e.g. {@code System.in}). The method
  * {@link #iterator()} returns an Iterator that blocks on calls to
  * {@link Iterator#hasNext()} until new classes are available, or no more
@@ -35,6 +51,13 @@ public class LazyTestsToRun extends TestsToRun {
 	private ClassLoader testClassLoader;
 	private PrintStream originalOutStream;
 
+	/**
+	 * C'tor
+	 * 
+	 * @param testSource source to read the tests from
+	 * @param testClassLoader class loader to load the test classes
+	 * @param originalOutStream the output stream to use when requesting new new tests
+	 */
 	public LazyTestsToRun(InputStream testSource, ClassLoader testClassLoader, PrintStream originalOutStream) {
 		super(Collections.emptyList());
 
@@ -104,6 +127,9 @@ public class LazyTestsToRun extends TestsToRun {
 
 	}
 
+	/* (non-Javadoc)
+	 * @see org.apache.maven.surefire.util.TestsToRun#iterator()
+	 */
 	public Iterator iterator() {
 		return new BlockingIterator();
 	}
@@ -122,6 +148,9 @@ public class LazyTestsToRun extends TestsToRun {
 		throw new UnsupportedOperationException("use method iterator()");
 	}
 
+	/* (non-Javadoc)
+	 * @see org.apache.maven.surefire.util.TestsToRun#toString()
+	 */
 	public String toString() {
 		StringBuffer sb = new StringBuffer("LazyTestsToRun ");
 		synchronized (workQueue) {
